Finalize moves and dependencies
Added missing modules, dependencies and imports.
