+++ bdssnmpadaptor/access.py
@@ -16,10 +16,10 @@ import time
 from bdssnmpadaptor.config import loadBdsSnmpAdapterConfigFile
 from bdssnmpadaptor.log import set_logging
 from bdssnmpadaptor.oidDb import OidDb
-from staticAndPredefinedOids import StaticAndPredefinedOids
-from mappingFuncModules.confd_local_system_software_info_confd import confd_local_system_software_info_confd
-from mappingFuncModules.confd_global_startup_status_confd import confd_global_startup_status_confd
-from mappingFuncModules.confd_global_interface_physical import confd_global_interface_physical
+from bdssnmpadaptor.predefined_oids import StaticAndPredefinedOids
+from bdssnmpadaptor.mapping_modules.confd_local_system_software_info_confd import confd_local_system_software_info_confd
+from bdssnmpadaptor.mapping_modules.confd_global_startup_status_confd import confd_global_startup_status_confd
+from bdssnmpadaptor.mapping_modules.confd_global_interface_physical import confd_global_interface_physical
 
 BIRTHDAY = time.time()
