+++ bdssnmpadaptor/mapping_modules/confd_global_startup_status_confd.py
