+++ bdssnmpadaptor/mapping_modules/lldpd_global_lldp_intf_status.py
@@ -1,12 +1,13 @@
 import os
 import sys
 sys.path.insert(0, os.path.abspath('..'))
-from bdssnmpadaptor.bdsMappingFunctions import bdsMappingFunctions
+from bdssnmpadaptor.mapping_functions import bdsMappingFunctions
 #import pytablewriter
 #import yaml
 from bdssnmpadaptor.oidDb import OidDbItem
 import struct
 import binascii
+import time
 
 
 IFTYPEMAP = {
@@ -22,78 +23,75 @@ IFOPERSTATUSMAP = {
 bigEndianFloatStruct = struct.Struct('>f')
 littleEndianShortStruct = struct.Struct('<h')
 IFMTU_LAMBDA = lambda x : int(littleEndianShortStruct.unpack(binascii.unhexlify(x))[0])
-IFSPEED_LAMBDA = lambda x : int(bigEndianFloatStruct.unpack(binascii.unhexlify(x))[0]/1000000*8)
+IFSPEED_LAMBDA = lambda x : int(bigEndianFloatStruct.unpack(binascii.unhexlify(x))[0]/1000*8)
 
 #HEX_STRING_LAMBDA = lambda x : int(x,16)
 #IFMTU_LAMBDA = lambda x : int("".join([m[2:4]+m[0:2] for m in [x[i:i+4] for i in range(0,len(x),4)]]),16)
 
-class confd_global_interface_container(object):
+class lldpd_global_lldp_intf_status (object):
 
     """
+        curl -X POST -H "Content-Type: application/json" -H "Accept: */*" -H "connection: close"\
+                       -H "Accept-Encoding: application/json"\
+                       -d '{"table": {"table_name": "global.lldp.intf.status"}}'\
+                       "http://10.0.3.10:2002/bds/table/walk?format=raw" | jq '.'
 
-    root@l2-pod2:~#     curl -X POST -H "Content-Type: application/json" -H "Accept: */*" -H "connection: close"\
-    >         -H "Accept-Encoding: application/json"\
-    >         -d '{"table": {"table_name": "global.interface.container"}}'\
-    >         "http://10.0.3.10:2002/bds/table/walk?format=raw" | jq '.'
-      % Total    % Received % Xferd  Average Speed   Time    Time     Time  Current
-                                     Dload  Upload   Total   Spent    Left  Speed
-    100 18629    0 18574  100    55  2826k   8570 --:--:-- --:--:-- --:--:-- 3023k
-    {
-      "table": {
-        "table_name": "global.interface.container"
-      },
-      "objects": [
         {
-          "sequence": 200197,
-          "update": true,
-          "attribute": {
-            "interface_name": "ifc-0/0/0/1",
-            "interface_description": "Container interface for ifp-0/0/1",
-            "encapsulation_type": "01",
-            "bandwidth": "4e9502f9",
-            "layer2_mtu": "f205",
-            "admin_status": "01",
-            "link_status": "01",
-            "mac_address": "b86a97a59201",
-            "physical_interfaces": [
-              "ifp-0/0/1"
-            ]
-          }
-        },
-        {
-          "sequence": 200198,
-          "update": true,
-          "attribute": {
-            "interface_name": "ifc-0/0/0/2",
-            "interface_description": "Container interface for ifp-0/0/2",
-            "encapsulation_type": "01",
-            "bandwidth": "4e9502f9",
-            "layer2_mtu": "f205",
-            "admin_status": "01",
-            "link_status": "01",
-            "mac_address": "b86a97a59202",
-            "physical_interfaces": [
-              "ifp-0/0/2"
-            ]
-          }
-        },
-        {
-          "sequence": 200061,
-          "update": true,
-          "attribute": {
-            "interface_name": "ifc-0/0/0/3",
-            "interface_description": "Container interface for ifp-0/0/3",
-            "encapsulation_type": "01",
-            "bandwidth": "4e9502f9",
-            "layer2_mtu": "f205",
-            "admin_status": "01",
-            "link_status": "01",
-            "mac_address": "b86a97a59203",
-            "physical_interfaces": [
-              "ifp-0/0/3"
-            ]
-          }
-        },
+          "objects": [
+            {
+              "attribute": {
+                "supported_breakout_modes": [
+                  "4x1G"
+                ],
+                "configured_layer2_mtu": "dc05",
+                "supported_max_layer2_mtu": "f205",
+                "supported_min_layer2_mtu": "4000",
+                "supported_auto_negotiation": "01",
+                "configured_bandwidth": "4e9502f9",
+                "default_bandwidth": "4e9502f9",
+                "supported_bandwidths": [
+                  "10.000 Gbps",
+                  "1.000 Gbps"
+                ],
+                "interface_name": "ifp-0/0/1",
+                "interface_description": "Physical interface #1 from node 0, chip 0",
+                "interface_type": "01",
+                "bandwidth": "4e9502f9",
+                "layer2_mtu": "f205",
+                "mac_address": "b86a97a59201",
+                "admin_status": "01",
+                "link_status": "01"
+              },
+              "update": true,
+              "sequence": 200197
+            },
+            {
+              "attribute": {
+                "supported_breakout_modes": [
+                  "4x1G"
+                ],
+                "configured_layer2_mtu": "dc05",
+                "supported_max_layer2_mtu": "f205",
+                "supported_min_layer2_mtu": "4000",
+                "supported_auto_negotiation": "01",
+                "configured_bandwidth": "4e9502f9",
+                "default_bandwidth": "4e9502f9",
+                "supported_bandwidths": [
+                  "10.000 Gbps",
+                  "1.000 Gbps"
+                ],
+                "interface_name": "ifp-0/0/2",
+                "interface_description": "Physical interface #2 from node 0, chip 0",
+                "interface_type": "01",
+                "bandwidth": "4e9502f9",
+                "layer2_mtu": "f205",
+                "mac_address": "b86a97a59202",
+                "admin_status": "01",
+                "link_status": "01"
+              },
+              "update": true,
+              "sequence": 200198
+            },
 
 
         5
@@ -123,110 +121,103 @@ class confd_global_interface_container(object):
         if str(newSequenceNumberList) == str(lastSequenceNumberList):
             pass #add logger statement
         else:
+            currentSysTime = int((time.time()-birthday)*100)
             targetOidDb.insertOid(newOidItem = OidDbItem(
-                bdsMappingFunc = "confd_global_interface_container",
+                bdsMappingFunc = "confd_global_interface_physical",
                 oid = "1.3.6.1.2.1.2.1.0",
                 name="ifIndex",
                 pysnmpBaseType="Integer32",
                 value=len(bdsJsonResponseDict["objects"])))
+            oidSegment = "1.3.6.1.2.1.2.2.1."
             targetOidDb.setLock()
             #targetOidDb.deleteOidsWithPrefix(oidSegment)  #delete existing TableOids
             for i,bdsJsonObject in enumerate(bdsJsonResponseDict["objects"]):
-                oidSegment = "1.3.6.1.2.1.2.2.1."
                 thisSequenceNumber = bdsJsonObject["sequence"]
                 ifName = bdsJsonObject["attribute"]["interface_name"]
                 index =  bdsMappingFunctions.ifIndexFromIfName(ifName)
                 #index =  i + 1
                 ifPhysicalLocation = bdsMappingFunctions.stripIfPrefixFromIfName(ifName)
                 targetOidDb.insertOid(newOidItem = OidDbItem(
-                    bdsMappingFunc = "confd_global_interface_container",
+                    bdsMappingFunc = "confd_global_interface_physical",
                     oid = oidSegment+"1."+str(index),
                     name="ifIndex",
                     pysnmpBaseType="Integer32",
                     value=int(index)))
                 targetOidDb.insertOid(newOidItem = OidDbItem(
-                    bdsMappingFunc = "confd_global_interface_container",
+                    bdsMappingFunc = "confd_global_interface_physical",
                     oid = oidSegment+"2."+str(index),
                     name="ifDescr",
                     pysnmpBaseType="OctetString",
                     value=ifPhysicalLocation))
                 targetOidDb.insertOid(newOidItem = OidDbItem(
-                    bdsMappingFunc = "confd_global_interface_container",
+                    bdsMappingFunc = "confd_global_interface_physical",
                     oid = oidSegment+"3."+str(index),
                     name="ifType",
                     pysnmpBaseType="Integer32",
-                    value=IFTYPEMAP[int(bdsJsonObject["attribute"]["encapsulation_type"])]))
+                    value=IFTYPEMAP[int(bdsJsonObject["attribute"]["interface_type"])]))
                 targetOidDb.insertOid(newOidItem = OidDbItem(
-                    bdsMappingFunc = "confd_global_interface_container",
+                    bdsMappingFunc = "confd_global_interface_physical",
                     oid = oidSegment+"4."+str(index),
                     name="ifMtu",
                     pysnmpBaseType="Integer32",
                     value=IFMTU_LAMBDA(bdsJsonObject["attribute"]["layer2_mtu"])))
                 targetOidDb.insertOid(newOidItem = OidDbItem(
-                    bdsMappingFunc = "confd_global_interface_container",
+                    bdsMappingFunc = "confd_global_interface_physical",
+                    oid = oidSegment+"5."+str(index),
+                    name="ifSpeed",
+                    pysnmpBaseType="Gauge32",
+                    value=IFSPEED_LAMBDA(bdsJsonObject["attribute"]["bandwidth"])))
+                targetOidDb.insertOid(newOidItem = OidDbItem(
+                    bdsMappingFunc = "confd_global_interface_physical",
                     oid = oidSegment+"6."+str(index),
                     name="ifPhysAddress",
                     pysnmpBaseType="OctetString",
                     pysnmpRepresentation="hexValue",
                     value=bdsJsonObject["attribute"]["mac_address"].replace(":","")))
                 targetOidDb.insertOid(newOidItem = OidDbItem(
-                    bdsMappingFunc = "confd_global_interface_container",
+                    bdsMappingFunc = "confd_global_interface_physical",
                     oid = oidSegment+"7."+str(index),
                     name="ifAdminStatus",
                     pysnmpBaseType="Integer32",
-                    value=bdsJsonObject["attribute"]["admin_status"]))
+                    value=IFOPERSTATUSMAP[int(bdsJsonObject["attribute"]["admin_status"])]))
                 targetOidDb.insertOid(newOidItem = OidDbItem(
-                    bdsMappingFunc = "confd_global_interface_container",
+                    bdsMappingFunc = "confd_global_interface_physical",
                     oid = oidSegment+"8."+str(index),
                     name="ifOperStatus",
                     pysnmpBaseType="Integer32",
                     value=IFOPERSTATUSMAP[int(bdsJsonObject["attribute"]["link_status"])]))
                 if len (lastSequenceNumberList) == 0:      #first run
                     targetOidDb.insertOid(newOidItem = OidDbItem(
-                        bdsMappingFunc = "confd_global_interface_container",
+                        bdsMappingFunc = "confd_global_interface_physical",
                         oid = oidSegment+"9."+str(index),
                         name="ifLastChange",
                         pysnmpBaseType="TimeTicks",
                         value=0 ))
-                elif thisSequenceNumber != lastSequenceNumberList[i] :    #status has changed
+                elif thisSequenceNumber != lastSequenceNumberList[i] :
                     targetOidDb.insertOid(newOidItem = OidDbItem(
-                        bdsMappingFunc = "confd_global_interface_container",
+                        bdsMappingFunc = "confd_global_interface_physical",
                         oid = oidSegment+"9."+str(index),
                         name="ifTableLastChange",
                         pysnmpBaseType="TimeTicks",
-                        value=int((time.time()-birthday)*100) ))
+                        value=currentSysTime ))
                 if len (lastSequenceNumberList) == 0:      #first run
                     targetOidDb.insertOid(newOidItem = OidDbItem(
-                        bdsMappingFunc = "confd_global_interface_container",
+                        bdsMappingFunc = "confd_global_interface_physical",
                         oid = "1.3.6.1.2.1.31.1.5",
                         name="ifTableLastChange",
                         pysnmpBaseType="TimeTicks",
                         value=0 ))
                     targetOidDb.insertOid(newOidItem = OidDbItem(
-                        bdsMappingFunc = "confd_global_interface_container",
+                        bdsMappingFunc = "confd_global_interface_physical",
                         oid = "1.3.6.1.2.1.31.1.6",
                         name="ifTableLastChange",
                         pysnmpBaseType="TimeTicks",
                         value=0 ))                  #Fixme - do we have to observe logicsl interfaces?
                 else:
                     targetOidDb.insertOid(newOidItem = OidDbItem(
-                        bdsMappingFunc = "confd_global_interface_container",
+                        bdsMappingFunc = "confd_global_interface_physical",
                         oid = "1.3.6.1.2.1.31.1.5",
                         name="ifTableLastChange",
                         pysnmpBaseType="TimeTicks",
-                        value=int((time.time()-birthday)*100) ))
-                #
-                # Change to ífXTable
-                #
-                oidSegment = "1.3.6.1.2.1.31.1.1.1."
-                targetOidDb.insertOid(newOidItem = OidDbItem(
-                    bdsMappingFunc = "confd_global_interface_container",
-                    oid = oidSegment+"15."+str(index),
-                    name="ifSpeed",
-                    pysnmpBaseType="Gauge32",
-                    value=IFSPEED_LAMBDA(bdsJsonObject["attribute"]["bandwidth"])))
-
-
-
-
+                        value=currentSysTime ))
             targetOidDb.releaseLock()
